fix(ci): clear lint debt + remove broken codecov upload
Two pre-existing issues surfaced by e18bc49 resolving CI workflow
paths:

1. Lint debt (27 violations across src/sonar + tests): pre-commit
   only-changed-files mode never caught these accumulating.
   Autofix applied for RUF046/TCH/UP/B; PLC0415 resolved case-by-case
   — lazy sonar.config import in daily_curves.py promoted to module
   top (db.session already forces Settings() instantiation upstream,
   so the lazy defer never bought anything).

2. Codecov v4 tokenless upload rejected; step removed. Coverage
   tracking continues via phase1-coverage-policy.md local gates.
   Codecov re-introduction deferred to Phase 2+ if needed.

Mypy remains clean. 202 unit tests continue to pass.

Completes CI remediation: 4820b85 (specs) → 0b2f43c (Python 3.12 +
FRED dummy) → e18bc49 (src-layout paths + repo rename) → this commit.

+++ src/sonar/connectors/te.py
@@ -130,7 +130,7 @@ class TEConnector:
                     country_code=country,
                     observation_date=obs_date,
                     tenor_years=10.0,
-                    yield_bps=int(round(close_pct * 100)),
+                    yield_bps=round(close_pct * 100),
                     source="TE",
                     source_series_id=symbol,
                 )
